Add files via upload
Final Project upload, I included the database information and screenshots of it running in my browser. I have been having some issues with the google map markers. But did solve the problem. The form is not finished, but I have built the backend, but still need to work on the scoring section of the code.